backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
langgraph
langchain-groq
pydantic
pydantic-settings
python-dotenv
alembic
python-multipart
